Update hello-world.md
I am suggesting an edit to the "Making and committing changes" section. The third instruction prompts the reader to use different "Markdown elements". Even after using the glossary, this instruction may not be very clear to a reader who is not familiar with writing code, using an IDE or GitHub. Keeping the instruction simple by asking the reader to "Try using different lines" may be more beginner friendly.